move get_real_source to Docsite
